+++ wefact/enums/__init__.py
@@ -1,2 +1,3 @@
 from .actions import Action
 from .invoice_actions import InvoiceAction
+from .interaction_actions import InteractionAction